backend/requirements.txt
# Updated requirements.txt for Python 3.13 compatibility
# Face Recognition App - macOS ARM64 Compatible

# Core Flask framework
Flask==2.3.2
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==22.0.0

# Computer Vision and Face Detection
opencv-python-headless==4.9.0.80
mtcnn==0.1.1
Pillow==10.4.0

# Deep Learning
tensorflow>=2.15.0,<2.17.0
keras-facenet==0.3.2

# Data Processing
numpy>=1.24.0,<2.0.0

matplotlib==3.7.2
scipy==1.15.3